fajita: Update libfod extension function prototypes

+++ fod/FodExtension.cpp
@@ -16,10 +16,10 @@
 
 #include <compositionengine/FodExtension.h>
 
-uint32_t getFodZOrder(uint32_t, bool touched) {
+uint64_t getFodZOrder(uint64_t, bool touched) {
     return touched ? 0xfc8 : 0xbd6;
 }
 
-uint32_t getFodUsageBits(uint32_t usageBits, bool) {
+uint64_t getFodUsageBits(uint64_t usageBits, bool) {
     return usageBits;
 }
